get_stats

Returns a count of sites, datasets, publications and other objects added to Neotoma during the requested time period.

get_stats(start = 0, end = 1, type = "dsdbmonth")

Arguments

start

The starting month (from present == 0) for which to generate the summary. Default is 0 (the current month).

end

The ending month (from present == 0) for which to generate the summary. Default is 1 (one month ago).

type

A character string indicating the type of summary to return. Options are dsdbmonth (the number of datasets in the Neotoma Database added per month), rawbymonth (the number of datasets, sites, publications, authors, countries and observations added per month), and dstypemonth (the number of datasets added per dataset type per month). Default is dsdbmonth.

Value

data.frame with summary statistics

Details

This function returns summaries about the data holdings within Neotoma using the existing Neotoma API's summary endpoint. This can provide information about recent uploads (the number of new sites uploaded within the last month, for example), or can be used to provide information about the overall number of sites/datasets (using an arbitrarily high value for end).

Author

Socorro Dominguez dominguezvid@wisc.edu

Simon Goring goring@wisc.edu

Examples

if (FALSE) { # \dontrun{
tryCatch({
last_month <- get_stats(start = 0, end = 1, type = "dsdbmonth")
}, error = function(e) {
 message("Neotoma server not responding. Try again later.")
})
} # }