get_taxon

a taxon object with detailed information

get_taxon(x = NA, ...)

# Default S3 method
get_taxon(...)

# S3 method for class 'numeric'
get_taxon(x, ...)

Arguments

x

taxon ID

...

accepted arguments, see details for more information.

Value

A Neotoma2 taxa object with datasets with the requested taxa.

Details

A taxa may have one or more taxa associated with it. The get_taxon function searches for taxa detail data within Neotoma while the get_taxa function searches for sites that contain the requested taxa. The function searches for each taxon by using a SQL query which accepts any one of the following parameters:

taxonid The unique taxon ID (integer) in Neotoma. Can be passed as a vector of taxa IDs.

taxonname Taxon name or partial name. You can pass wildcard characters using %.

taxagroup The taxonomic grouping the taxon belongs to, from the Neotoma taxagrouptypes table.

ecolgroup The ecological group of interest for the Neotoma taxon (from the Neotoma ecolgrouptypes table)

status The taxonomic extinction status of the taxon, either extinct (1, True) or extant (0, False).

Author

Socorro Dominguez dominguezvid@wisc.edu