set chronology information for a new record.

Create a new chronology for a record. Within Neotoma all chronologies have unique numeric identifiers. Within R, because of the need to use the identifiers across objects, and because we want to avoid conflicts between naming systems, a universally unique identifier (UUID) is created for the object ID.

Function to create new chronology objects for personal analysis. The new object will not be uploaded to the database.

set_chronology(
  x = NA,
  chronologyid = NA_integer_,
  notes = NA_character_,
  contact = list(),
  agemodel = NA_character_,
  ageboundolder = NA_integer_,
  ageboundyounger = NA_integer_,
  isdefault = NA_integer_,
  dateprepared = as.Date(character(0)),
  modelagetype = NA_character_,
  chronologyname = NA_character_,
  chroncontrols = data.frame(0)
)

Arguments

x

Object to be set as a chronology

chronologyid

An optional value. Will be assigned a unique identifier if not provided.

notes

Additional notes about the chronology. For more modern models, often the function call to Bacon or Bchron is added here.

contact

A contacts object, identifying the individual(s) who created the chronology

agemodel

A string representing the age model name, for example "Crummy linear interpolation".

ageboundolder

The ageboundolder is assigned the oldest sample age rounded up to the nearest 10

ageboundyounger

The ageboundyounger is assigned the oldest sample age rounded up to the nearest 10

isdefault

Defines whether the model is the default for the collection unit for a particular model age type.

dateprepared

The date at which the age model was prepared.

modelagetype

The age type for the model. For validation, the models should be one of the valid Neotoma agetypes: https://api.neotomadb.org/v2.0/data/dbtables?table=agetypes

chronologyname

A valid name for the chronology.

chroncontrols

A data.frame containing the chronological controls for the age model.

Value

chronology object

Author

Socorro Dominguez dominguezvid@wisc.edu

Simon Goring goring@wisc.edu