set Site Information for Fossil Sites

Function to create new collection unit objects for personal analysis. The new object will not be uploaded to the database.

set_collunit(
  x = NA,
  collectionunitid = NA_integer_,
  notes = NA_character_,
  handle = NA_character_,
  colldate = as.Date(character(1)),
  location = NA_character_,
  waterdepth = NA_integer_,
  gpslocation = st_as_sf(st_sfc()),
  collunittype = NA_character_,
  collectiondevice = NA_character_,
  collectionunitname = NA_character_,
  depositionalenvironment = NA_character_,
  datasets = NULL,
  chronologies = NULL,
  defaultchronology = NA_integer_,
  speleothems = NULL
)

Arguments

x

object to be set as collunit

collectionunitid

collection unit identifier

notes

notes

handle

handle

colldate

collection date

location

location of the collection unit

waterdepth

depth at where the sample is taken

gpslocation

location with GPS

collunittype

type of collection unit

collectiondevice

device used to collect the sample

collectionunitname

name of the collection unit

depositionalenvironment

depositional environment

datasets

datasets that the collection unit has

chronologies

chronologies taken from the collection unit

defaultchronology

best chronology model identifier to be used with this collection unit

speleothems

speleothems associated with the collection unit

Value

collunit object

Author

Socorro Dominguez dominguezvid@wisc.edu

Simon Goring goring@wisc.edu

Examples

{
# Create a collunit
my_collunit <- set_collunit(notes = "my lake")
}