Set contact information for a new record.

Create a new contact for a record. Within Neotoma all chronologies have unique numeric identifiers. Within R, because of the need to use the indentifiers across objects, and because we want to avoid conflicts between naming systems, a universally unique identifier (UUID) is created for the object ID.This is not to be updated to the database.

set_contact(
  x = NA,
  contactid = NA_integer_,
  familyname = NA_character_,
  leadinginitials = NA_character_,
  givennames = NA_character_,
  contactname = NA_character_,
  suffix = NA_character_,
  ORCID = NA_character_,
  title = NA_character_,
  institution = NA_character_,
  email = NA_character_,
  phone = NA_character_,
  contactstatus = NA_character_,
  fax = NA_character_,
  url = NA_character_,
  address = NA_character_,
  notes = NA_character_
)

Arguments

x

Object to be set as a contact

contactid

An arbitrary Contact identification number.

familyname

Family or surname name of a person.

leadinginitials

Leading initials for given or forenames without spaces (e.g. G.G.).

givennames

Given or forenames of a person (e.g. George Gaylord). Initials with spaces are used if full given names are not known (e.g. G. G).

contactname

Full name of a person (e.g. George Gaylord Simpson).

suffix

Suffix of a person’s name (e.g. Jr., III).

ORCID

A unique ORCID (see https://orcid.org).

title

A person’s title (e.g. Dr., Prof., Prof. Dr).

institution

The institution where an individual works.

email

An individuals email address

phone

Their phone number

contactstatus

Are they "active" or "retired"?

fax

Do people still use fax machines?

url

Their homepage

address

A physical address

notes

Notes about the individual

Value

contact object