set Site Information for Fossil Sites

set Site Information for Fossil Sites

set_site(
  x = NA,
  siteid = NA_integer_,
  sitename = NA_character_,
  geography = sf::st_as_sf(sf::st_sfc()),
  altitude = NA_integer_,
  geopolitical = list(),
  area = NA_integer_,
  notes = NA_character_,
  description = NA_character_,
  collunits = NULL
)

Arguments

x

Object to be set as a site

siteid

The unique site id for a site. If this site is new to Neotoma then leave the ID as NA.

sitename

Name of a site.

geography

A sf object representing the site location, either as a polygon or point.

altitude

altitude/elevation of the site.

geopolitical

The geopolitical unit in which the site is located.

area

The area of the site or depositional basin in ha. Can be calculated from the polygon.

notes

additional information of the site

description

Function to create new site objects. The new object will not be uploaded to the database.

collunits

Collection units in the site

Value

site object

Author

Socorro Dominguez dominguezvid@wisc.edu

Examples

{
# Create a site called "My Lake", to
x <- sf::st_as_sf(sf::st_sfc(sf::st_point(c(5,5))))
my_site <- set_site(sitename = "My Lake",
                    geography = x,
                    description = "my lake",
                    altitude = 30)
}