set Site Information for Fossil Sites

set Site Information for Fossil Sites

set_speleothem(
  x = NA,
  entityid = NA_integer_,
  entityname = NA_character_,
  siteid = NA_integer_,
  collectionunitid = NA_integer_,
  dripheight = NA_integer_,
  monitoring = FALSE,
  geology = NA_character_,
  relativeage = NA_character_,
  speleothemtype = NA_character_,
  dripheightunits = NA_character_,
  entitycovertype = NA_character_,
  entrancedistance = NA_integer_,
  landusecovertype = NA_character_,
  speleothemdriptype = NA_character_,
  landusecoverpercent = NA_integer_,
  vegetationcovertype = NA_character_,
  entitycoverthickness = NA_integer_,
  entrancedistanceunits = NA_character_,
  vegetationcoverpercent = NA_integer_
)

Arguments

x

Object to be set as a speleothem

siteid

The unique site id for a site. If this site is new to Neotoma then leave the ID as NA (the default).

sitename

Actual site name as a character string.

geography

An sf object representing the site location, either as a polygon or point.

altitude

altitude/elevation of the site.

geopolitical

The geopolitical unit in which the site is located.

area

The area of the site or depositional basin in ha. Can be calculated from the polygon.

description

Function to create new site objects for personal analysis. The new object will not be uploaded to the database.

notes

additional information of the site

collunits

Collection units in the site

Value

site object